pyproject.toml
[project]
name = "weather-etl-pipeline"
version = "0.1.0"
description = "Weather ETL Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flake8>=7.3.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["setuptools", "wheel"]

[tool.flake8]
max-line-length = 100
ignore = "E203, W503, E501, F401"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    "migrations",
    "venv",
    "env",
    ".env"
]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:S101"
]
format = "pylint"
